* Separating evolve API for Python and C++ * Adding CUDAQ DYNAMICS target check * Formatting
